refactor: move quarter-revenue vocabulary to @sznyt/shared
The QuarterRevenue response shape crosses the API boundary, so per
ADR-0002 it lives in the shared workspace package instead of being
hand-copied in the banner component.
