+++ backend/revenue/index.ts
@@ -1,7 +1,5 @@
 export {
   computeQuarterRevenue,
   QUARTERLY_REVENUE_CAP_PLN,
-  type QuarterRevenue,
-  type QuarterRevenueThreshold,
   type RevenueOrder,
 } from "./computeQuarterRevenue.ts";
