+++ packages/shared/src/index.ts
@@ -16,3 +16,5 @@ export {
 } from "./shipping.ts";
 export type { ShippingMethod } from "./shipping.ts";
 export type { ShippingAddress } from "./shippingAddress.ts";
+export { REVENUE_THRESHOLDS } from "./revenue.ts";
+export type { QuarterRevenue, QuarterRevenueThreshold } from "./revenue.ts";
